Attempting to use MySQL database

+++ Migrate/Program.cs
@@ -20,7 +20,9 @@ foreach (var filename in removeFileList.Where(File.Exists))
 }
 
 // Database
-var freeSql = FreeSqlFactory.Create("Data Source=app.db;Synchronous=Off;Cache Size=5000;");
+// var freeSql = FreeSqlFactory.Create("Data Source=app.db;Synchronous=Off;Cache Size=5000;");
+var freeSql = FreeSqlFactory.CreateMySql(
+    "Data Source=sbtimi.com;Port=3306;User ID=root;Password=6Mynaom7Eejj0U0fqoqh; Initial Catalog=starblog;Charset=utf8mb4; SslMode=none;Min pool size=1");
 var postRepo = freeSql.GetRepository<Post>();
 var categoryRepo = freeSql.GetRepository<Category>();
 
@@ -29,8 +31,11 @@ WalkDirectoryTree(new DirectoryInfo(importDir));
 
 // Overwrite database
 var destFile = Path.GetFullPath("../Web/app.db");
-Console.WriteLine($"Overwrite {destFile}");
-File.Copy("app.db", destFile, true);
+if (File.Exists("app.db") && File.Exists(destFile))
+{
+    Console.WriteLine($"覆盖数据库：{destFile}");
+    File.Copy("app.db", destFile, true);
+}
 
 void WalkDirectoryTree(DirectoryInfo root)
 {
